refactor(notify): move notify from plugin to lib
The notify renderer is infrastructure consumed by every plugin's
`ttymap.notify(msg)` call, not a user-facing feature like wiki or
aircraft. Its knobs (TTL, ring cap, max text width) belong on a
caller-provided `setup({...})` rather than baked locals.

  runtime/plugin/notify.lua  → runtime/lua/ttymap/notify.lua
  (auto-activate)              (explicit `require("ttymap.notify").setup()`)

Bundled `runtime/init.lua` calls `setup()` with no args (defaults
match the previous baked constants). Users wanting different
knobs:

  require("ttymap.notify").setup({ ttl_s = 8, ring_cap = 16 })

Skipping the call disables the toast renderer entirely; the
`Event::Notify` bus event still fires for any other subscriber.

`setup()` is idempotent — a second call logs a warning and returns
without double-subscribing.

README + docs/lua-architecture.md updated: bundled-plugin count
17 → 16, notify row dropped from the table, notify section
rewritten to describe the lib + setup contract.
